V3.2 with change from current to stable/unstable
Version 3.2 is what was in the ‘current’ directory before.  If you want
to continue to use version 3.2, you will need to either directly run
from that directory, or point to the unstable link in the GTDef
directory before.  After looking at some issues with editing from
multiple locations, I decided that this was a better structure.